Created GameClass and SceneManager Class
Added the GameClass, which will hold everything about the logic of the game. Also added the SceneManager Class: This class is responsible that the correct Scene is show/Updated/Switched/Player gets switched. In this current commit, The GraphicsView and CurrentScene are still not connected, And some functionality with graphics scene needs to be implemented, but this is practicly done. However to check that it does behave correctly a few print slots have been placed that can switch.