Added Projectiles/Bullets
Added Projectiles/Bullets, they behave correctly eg. have a limited flight time, delete them self correctly after that flight time or delete them self when they hit a wall. For this a new slot was made in Room. As also the Init of SceneObject was changed to now for all SceneObjects have the variable current room.